requirements.txt
# Package and package dependencies
-e .

# Development dependencies
setuptools>=67.6.1,<67.7
wheel>=0.40.0,<0.41
twine>=4.0.2,<4.1
coverage>=7.2.3,<7.3

# Linters and formatters
pre-commit>=3.2.2,<3.3

# TestApp dependencies
django>=3.2,<4.3
djangorestframework>=3.14.0,<4
